Move common utility methods for conversion to package level
