+++ converter/ena/ena_study.py
@@ -2,6 +2,7 @@ from xml.etree.ElementTree import Element
 
 from lxml import etree
 
+from converter import get_value_by_key_path, fixed_attribute
 from converter.ena.base import BaseEnaConverter
 
 ADDITIONAL_ATTRIBUTE_KEYS = ['uuid.uuid', 'content.project_core.project_short_name', 'technology']
@@ -11,11 +12,11 @@ class EnaStudyConverter(BaseEnaConverter):
 
     def __init__(self):
         study_spec = {
-            '@center_name': ['', BaseEnaConverter._fixed_attribute, 'HCA'],
+            '@center_name': ['', fixed_attribute, 'HCA'],
             'DESCRIPTOR': {
                 'STUDY_TITLE': ['project_title'],
                 'STUDY_TYPE': {
-                    '@existing_study_type': ['', BaseEnaConverter._fixed_attribute, 'Transcriptome Analysis']
+                    '@existing_study_type': ['', fixed_attribute, 'Transcriptome Analysis']
                 },
                 'STUDY_DESCRIPTION': ['project_description']
             }
@@ -27,7 +28,7 @@ class EnaStudyConverter(BaseEnaConverter):
         attributes = etree.SubElement(xml_element, 'STUDY_ATTRIBUTES')
         for key in ADDITIONAL_ATTRIBUTE_KEYS:
             key_list: list = key.split('.')
-            value: str = EnaStudyConverter._get_value_by_key_path(entity, key_list)
+            value: str = get_value_by_key_path(entity, key_list)
             if value:
                 BaseEnaConverter._make_attribute(
                     attributes, 'STUDY_ATTRIBUTE', key_list[-1], value)
